requirements.txt
psycopg2-binary
python-dotenv
# Core dependencies
pyyaml
# Databricks connector for table size calculation
databricks-sql-connector

fastapi
uvicorn[standard]
pydantic
pyyaml
python-multipart
psycopg2-binary

# Concurrency testing dependencies
databricks-sdk
asyncpg
sqlalchemy
greenlet